Added cli for cascadia
