+++ casanovo_cascadia/denovo/model_runner.py
@@ -48,6 +48,8 @@ class ModelRunner:
     overwrite_ckpt_check: bool, optional
         Whether to check output_dir (if not `None`) for conflicting
         checkpoint files.
+    casanovo: bool, optional
+        Whether or not to load data for the Casanovo or Cascadia models
     """
 
     def __init__(
@@ -57,6 +59,7 @@ class ModelRunner:
         output_dir: Optional[Path | None] = None,
         output_rootname: Optional[str | None] = None,
         overwrite_ckpt_check: Optional[bool] = True,
+        casanovo: Optional[bool] = True,
     ) -> None:
         """Initialize a ModelRunner."""
         self.config = config
@@ -64,6 +67,7 @@ class ModelRunner:
         self.output_dir = output_dir
         self.output_rootname = output_rootname
         self.overwrite_ckpt_check = overwrite_ckpt_check
+        self.casanovo = casanovo
 
         # Initialized later.
         self.tmp_dir = None
@@ -167,7 +171,7 @@ class ModelRunner:
         self.writer.set_ms_run(test_paths)
         self.initialize_data_module(test_paths=test_paths)
         self.loaders.protein_database = self.model.protein_database
-        self.loaders.setup(stage="test", annotated=False)
+        self.loaders.setup(stage="test", annotated=False, casanovo=True)
         self.trainer.predict(self.model, self.loaders.db_dataloader())
 
     def train(
@@ -647,6 +651,7 @@ class ModelRunner:
             shuffle_buffer_size=self.config.shuffle_buffer_size,
             n_workers=self.config.n_workers,
             lance_dir=lance_dir,
+            casanovo=self.casanovo,
         )
 
     @staticmethod
