+++ casanovo_cascadia/shared_loading.py
@@ -1,0 +1,579 @@
+import rich_click as click
+from typing import Optional, Tuple, List
+import re
+from pathlib import Path
+import os
+import hashlib
+import urllib.parse
+import requests
+import datetime
+import tqdm
+import functools
+import shutil
+import logging
+import utils
+import github
+import sys
+
+logger = logging.getLogger(__name__)
+
+
+class _SharedFileIOParams(click.RichCommand):
+    """File IO options shared between most Casanovo commands"""
+
+    def __init__(self, *args, **kwargs) -> None:
+        """Define shared options."""
+        super().__init__(*args, **kwargs)
+        self.params += [
+            click.Option(
+                ("-d", "--output_dir"),
+                help="The destination directory for output files.",
+                type=click.Path(dir_okay=True),
+            ),
+            click.Option(
+                ("-o", "--output_root"),
+                help="The root name for all output files.",
+                type=str,
+            ),
+            click.Option(
+                ("-f", "--force_overwrite"),
+                help="Whether to overwrite output files.",
+                is_flag=True,
+                show_default=True,
+                default=False,
+            ),
+            click.Option(
+                ("-v", "--verbosity"),
+                help=(
+                    "Set the verbosity of console logging messages."
+                    " Log files are always set to 'debug'."
+                ),
+                type=click.Choice(
+                    ["debug", "info", "warning", "error"],
+                    case_sensitive=False,
+                ),
+                default="info",
+            ),
+        ]
+
+
+class _SharedParams(_SharedFileIOParams):
+    """Options shared between main Casanovo commands"""
+
+    def __init__(self, *args, **kwargs) -> None:
+        """Define shared options."""
+        super().__init__(*args, **kwargs)
+        self.params += [
+            click.Option(
+                ("-m", "--model"),
+                help="""Either the model weights (.ckpt file) or a URL pointing to 
+                the model weights file. If not provided, Casanovo will try to 
+                download the latest release automatically.""",
+            ),
+            click.Option(
+                ("-c", "--config"),
+                help="The YAML configuration file overriding the default options.",
+                type=click.Path(exists=True, dir_okay=False),
+            ),
+        ]
+
+
+def _setup_output(
+    output_dir: str | None,
+    output_root: str | None,
+    overwrite: bool,
+    verbosity: str,
+    model_name: str,
+) -> Tuple[Path, str]:
+    """
+    Set up the output directory, output file root name, and logging.
+
+    Parameters:
+    -----------
+    output_dir : str | None
+        The path to the output directory. If `None`, the output
+        directory will be resolved to the current working directory.
+    output_root : str | None
+        The base name for the output files. If `None` the output root
+        name will be resolved to casanovo_<current date and time>
+    overwrite: bool
+        Whether to overwrite log file if it already exists in the output
+        directory.
+    verbosity : str
+        The verbosity level for logging.
+
+    Returns:
+    --------
+    Tuple[Path, str]
+        A tuple containing the resolved output directory and root name
+        for output files.
+    """
+    if output_root is None:
+        output_root = (
+            f"{model_name}_{datetime.datetime.now().strftime('%Y%m%d%H%M%S')}"
+        )
+
+    if output_dir is None:
+        output_path = Path.cwd()
+    else:
+        output_path = Path(output_dir).expanduser().resolve()
+        if not output_path.is_dir():
+            output_path.mkdir(parents=True)
+            logger.warning(
+                "Target output directory %s does not exists, so it will be "
+                "created.",
+                output_path,
+            )
+
+    if not overwrite:
+        utils.check_dir_file_exists(output_path, f"{output_root}.log")
+
+    log_file_path = output_path / f"{output_root}.log"
+    setup_logging(log_file_path, verbosity)
+    return output_path, output_root
+
+
+def setup_logging(
+    log_file_path: Path,
+    verbosity: str,
+) -> None:
+    """
+    Set up the logger.
+
+    Logging occurs to the command-line and to the given log file.
+
+    Parameters
+    ----------
+    log_file_path: Path
+        The log file path.
+    verbosity : str
+        The logging level to use in the console.
+    """
+    logging_levels = {
+        "debug": logging.DEBUG,
+        "info": logging.INFO,
+        "warning": logging.WARNING,
+        "error": logging.ERROR,
+    }
+
+    # Configure logging.
+    logging.captureWarnings(True)
+    root_logger = logging.getLogger()
+    root_logger.setLevel(logging.DEBUG)
+    warnings_logger = logging.getLogger("py.warnings")
+
+    # Formatters for file vs console:
+    console_formatter = logging.Formatter("{levelname}: {message}", style="{")
+    log_formatter = logging.Formatter(
+        "{asctime} {levelname} [{name}/{processName}] {module}.{funcName} : "
+        "{message}",
+        style="{",
+    )
+
+    console_handler = logging.StreamHandler(sys.stderr)
+    console_handler.setLevel(logging_levels[verbosity.lower()])
+    console_handler.setFormatter(console_formatter)
+    root_logger.addHandler(console_handler)
+    warnings_logger.addHandler(console_handler)
+    file_handler = logging.FileHandler(log_file_path, encoding="utf8")
+    file_handler.setFormatter(log_formatter)
+    root_logger.addHandler(file_handler)
+    warnings_logger.addHandler(file_handler)
+
+    # Disable dependency non-critical log messages.
+    logging.getLogger("depthcharge").setLevel(
+        logging_levels[verbosity.lower()]
+    )
+    logging.getLogger("fsspec").setLevel(logging.WARNING)
+    logging.getLogger("github").setLevel(logging.WARNING)
+    logging.getLogger("h5py").setLevel(logging.WARNING)
+    logging.getLogger("numba").setLevel(logging.WARNING)
+    logging.getLogger("pytorch_lightning").setLevel(logging.WARNING)
+    logging.getLogger("torch").setLevel(logging.WARNING)
+    logging.getLogger("urllib3").setLevel(logging.WARNING)
+
+
+_DEFAULT_MODEL_ID = "orbitrap"
+
+
+def _normalize(s: str) -> str:
+    """
+    Normalizes model selector.
+
+    Removes hyphens, underscores, and whitespace characters.
+
+    Parameters
+    ----------
+    s: String
+       The string to be normalized.
+
+    Returns
+    -------
+    String
+       The normalized model selector.
+    """
+    return re.sub(r"[-_\s]", "", s).lower()
+
+
+def _resolve_selector(selector: str, candidates: List[str]) -> str:
+    """
+    Resolves the model selector to a present model.
+
+    Resolve a model selector to a canonical model ID via:
+      1. Exact normalized match
+      2. Unique normalized prefix match
+      3. Unique normalized substring match
+
+    Parameters
+    ----------
+    selector: String
+       The parsed selector from _parse_ckpt.
+    candidates: List[str]
+        The current models present in the cache.
+
+    Returns
+    -------
+    String
+        Best matching model to selector
+    """
+    norm = _normalize(selector)
+    norm_map = {c: _normalize(c) for c in candidates}
+
+    for matches in (
+        [c for c, nc in norm_map.items() if nc == norm],
+        [c for c, nc in norm_map.items() if nc.startswith(norm)],
+        [c for c, nc in norm_map.items() if norm in nc],
+    ):
+        if len(matches) == 1:
+            return matches[0]
+        if len(matches) > 1:
+            options = ", ".join(sorted(matches))
+            raise ValueError(
+                f"Ambiguous model selector '{selector}'. Matching models:\n"
+                + f"{options} \n Please specify one of these options."
+            )
+
+    available = "\n".join(f"  {c}" for c in sorted(candidates))
+    raise ValueError(
+        f"Unknown model selector '{selector}'.\nAvailable models:\n{available}"
+    )
+
+
+def _get_model_weights(
+    selector: str,
+    cache_dir: Path,
+    casanovo_version: Tuple[int, int, int],
+    ckpt_regex: re.Pattern,
+) -> Path:
+    """
+    Use cached model weights or download them from GitHub.
+
+    If no weights file (extension: .ckpt) is available in the cache
+    directory, it will be downloaded from a release asset on GitHub.
+    Model weights are retrieved by matching release version and by model selector.
+    The model selector indicates the type of model to be used (e.g., orbitrap, timstof).
+    Partial (or full) case-insensitive and character stripped comparisons are made
+    to the existing models to find the closest version.
+    If no model weights for an identical release (major, minor, patch), alternative
+    releases with matching (i) major and minor, or (ii) major versions
+    will be used. If no matching release can be found, no model weights
+    will be downloaded.
+
+    Note that the GitHub API is limited to 60 requests from the same IP
+    per hour.
+
+    Parameters
+    ----------
+    cache_dir : Path
+        Model weights cache directory path.
+    selector : String
+        The model name (e.g., timstof, orbitrap)
+
+    Returns
+    -------
+    Path
+        The path of the model weights file.
+    """
+    os.makedirs(cache_dir, exist_ok=True)
+
+    # Parse filenames into (model_id, path, version) triples.
+    local = []
+    for fn in os.listdir(cache_dir):
+        parsed = _parse_ckpt(fn, ckpt_regex)
+        if parsed:
+            model_id, version = parsed
+            path = (cache_dir / fn) if cache_dir else Path(fn)
+            local.append((model_id, path, version))
+
+    if local:
+        try:
+            canonical_id = _resolve_selector(
+                selector, list({mid for mid, _, _ in local})
+            )
+        except ValueError as exc:
+            if "Unknown model selector" not in str(exc):
+                raise
+            canonical_id = None
+
+        if canonical_id is not None:
+            family = [(p, v) for mid, p, v in local if mid == canonical_id]
+            best = _best_ckpt(family, casanovo_version)
+            if best:
+                logger.info(
+                    "Model weights file %s retrieved from local cache", best
+                )
+                return best
+
+    repo = github.Github().get_repo("Noble-Lab/casanovo")
+
+    github_ckpts = []
+    for release in repo.get_releases():
+        for asset in release.get_assets():
+            parsed = _parse_ckpt(asset.name, ckpt_regex)
+            if parsed:
+                model_id, version = parsed
+                github_ckpts.append(
+                    (model_id, asset.name, version, asset.browser_download_url)
+                )
+
+    if not github_ckpts:
+        raise ValueError(
+            "No canonical model checkpoints found on GitHub. "
+            "Specify weights explicitly with '--model'."
+        )
+
+    canonical_id = _resolve_selector(
+        selector, list({mid for mid, _, _, _ in github_ckpts})
+    )
+    family_gh = [
+        (Path(fn), v, url)
+        for mid, fn, v, url in github_ckpts
+        if mid == canonical_id
+    ]
+    best = _best_ckpt([(p, v) for p, v, _ in family_gh], casanovo_version)
+
+    if best is None:
+        available = "\n".join(
+            f"  {p.name}" for p, _, _ in sorted(family_gh, key=lambda x: x[1])
+        )
+        raise ValueError(
+            f"No compatible '{canonical_id}' checkpoint found for Casanovo "
+            f"{'.'.join(map(str, casanovo_version))}.\n"
+            f"Available {canonical_id} checkpoints:\n{available}\n"
+            "Please upgrade Casanovo, choose another model, or provide a "
+            "local path or URL with '--model'."
+        )
+
+    url = next(url for p, _, url in family_gh if p == best)
+    dest = cache_dir / best.name
+    _download_weights(url, dest)
+    return dest
+
+
+def _parse_ckpt(
+    filename: str, ckpt_regex: re.Pattern
+) -> Optional[Tuple[str, Tuple[int, int, int]]]:
+    """
+    Parses the checkpoint filename.
+
+    Extracts the major, minor model versions and model selector.
+
+    Parameters
+    ----------
+    filename: String
+        The name of the model file provided.
+
+    Returns
+    -------
+    Optional[Tuple[str, Tuple[int, int, int]]]
+        Method could return None if filename does not match _CKPT_RE
+        Returns a Tuple with the String set to the model selector,
+        and the integer tuple values set to the Casanovo compatibility
+        version.
+    """
+    parsed_string = ckpt_regex.match(os.path.basename(filename).lower())
+    if not parsed_string:
+        return None
+    return parsed_string.group(1), (
+        int(parsed_string.group(2)),
+        int(parsed_string.group(3)),
+        int(parsed_string.group(4)),
+    )
+
+
+def _best_ckpt(
+    checkpoints: List[Tuple[Path, Tuple[int, int, int]]],
+    version: Tuple[int, int, int],
+) -> Optional[Path]:
+    """
+    Pick best checkpoint by: exact patch > latest same minor > latest same major.
+
+    Parameters
+    ----------
+    checkpoints: List[Tuple[Path, Tuple[int, int, int]]]
+        Possible checkpoints.
+    version: Tuple[int, int, int]
+        Intended version.
+    """
+    maj, min_, pat = version
+    for subset in (
+        [x for x in checkpoints if x[1] == (maj, min_, pat)],
+        [x for x in checkpoints if x[1][:2] == (maj, min_)],
+        [x for x in checkpoints if x[1][0] == maj],
+    ):
+        if subset:
+            return max(subset, key=lambda x: x[1])[0]
+    return None
+
+
+def _is_valid_url(file_url: str) -> bool:
+    """
+    Determine whether file URL is a valid URL
+
+    Parameters
+    ----------
+    file_url : str
+        url to verify
+
+    Return
+    ------
+    is_url : bool
+        whether file_url is a valid url
+    """
+    try:
+        result = urllib.parse.urlparse(file_url)
+        return all([result.scheme, result.netloc])
+    except ValueError:
+        return False
+
+
+def _get_weights_from_url(
+    file_url: str,
+    cache_dir: Path,
+    force_download: Optional[bool] = False,
+) -> Path:
+    """
+    Resolve weight file from URL
+
+    Attempt to download weight file from URL if weights are not already
+    cached - otherwise use cached weights. Downloaded weight files will
+    be cached.
+
+    Parameters
+    ----------
+    file_url : str
+        URL pointing to model weights file.
+    cache_dir : Path
+        Model weights cache directory path.
+    force_download : Optional[bool], default=False
+        If True, forces a new download of the weight file even if it
+        exists in the cache.
+
+    Returns
+    -------
+    Path
+        Path to the cached weights file.
+    """
+    if not _is_valid_url(file_url):
+        raise ValueError("file_url must point to a valid URL")
+
+    os.makedirs(cache_dir, exist_ok=True)
+    cache_file_name = Path(urllib.parse.urlparse(file_url).path).name
+    url_hash = hashlib.shake_256(file_url.encode("utf-8")).hexdigest(5)
+    cache_file_dir = cache_dir / url_hash
+    cache_file_path = cache_file_dir / cache_file_name
+
+    if cache_file_path.is_file() and not force_download:
+        cache_time = cache_file_path.stat()
+        url_last_modified = 0
+
+        try:
+            file_response = requests.head(file_url)
+            if file_response.ok:
+                if "Last-Modified" in file_response.headers:
+                    url_last_modified = datetime.datetime.strptime(
+                        file_response.headers["Last-Modified"],
+                        "%a, %d %b %Y %H:%M:%S %Z",
+                    ).timestamp()
+            else:
+                logger.warning(
+                    "Attempted HEAD request to %s yielded non-ok status code—"
+                    "using cached file",
+                    file_url,
+                )
+        except (
+            requests.ConnectionError,
+            requests.Timeout,
+            requests.TooManyRedirects,
+        ):
+            logger.warning(
+                "Failed to reach %s to get remote last modified time—using "
+                "cached file",
+                file_url,
+            )
+
+        if cache_time.st_mtime > url_last_modified:
+            logger.info(
+                "Model weights %s retrieved from local cache", file_url
+            )
+            return cache_file_path
+
+    _download_weights(file_url, cache_file_path)
+    return cache_file_path
+
+
+def _download_weights(file_url: str, download_path: Path) -> None:
+    """
+    Download weights file from URL.
+
+    Download the model weights file from the specified URL and save it
+    to the given path. Ensures the download directory exists, and uses a
+    progress
+    bar to indicate download status.
+
+    Parameters
+    ----------
+    file_url : str
+        URL pointing to the model weights file.
+    download_path : Path
+        Path where the downloaded weights file will be saved.
+    """
+    download_file_dir = download_path.parent
+    os.makedirs(download_file_dir, exist_ok=True)
+    response = requests.get(file_url, stream=True, allow_redirects=True)
+    response.raise_for_status()
+    file_size = int(response.headers.get("Content-Length", 0))
+    desc = "(Unknown total file size)" if file_size == 0 else ""
+    response.raw.read = functools.partial(
+        response.raw.read, decode_content=True
+    )
+
+    with (
+        tqdm.tqdm.wrapattr(
+            response.raw, "read", total=file_size, desc=desc
+        ) as r_raw,
+        open(download_path, "wb") as file,
+    ):
+        shutil.copyfileobj(r_raw, file)
+
+
+def split_version(version: str) -> Tuple[str, str, str]:
+    """
+    Split the version into its semantic versioning components.
+
+    Parameters
+    ----------
+    version : str
+        The version number.
+
+    Returns
+    -------
+    major : str
+        The major release.
+    minor : str
+        The minor release.
+    patch : str
+        The patch release.
+    """
+    version_regex = re.compile(r"(\d+)\.(\d+)\.*(\d*)(?:.dev\d+.+)?")
+    return tuple(g for g in version_regex.match(version).groups())
