+++ tests/unit_tests/test_config.py
@@ -3,7 +3,7 @@
 import pytest
 import yaml
 
-from casanovo.config import Config
+from casanovo_cascadia.config import Config
 
 
 def test_default():
